+++ lib/service/api/apiAccount.dart
@@ -512,10 +512,9 @@ $validUntil: UInt32,$scalar: String!, $field: String!) {
       String pubKey = acc['pubKey'];
       store.walletConnectService?.emitAccountsChanged(pubKey);
       webApi.assets.fetchAllTokenAssets();
-      webApi.assets.fetchTransactions(pubKey);
       webApi.assets.fetchPendingTransactions(pubKey);
       webApi.assets.fetchPendingZkTransactions(pubKey);
-      webApi.assets.fetchZkTransactions(pubKey);
+      webApi.assets.fetchFullTransactions(pubKey);
       return true;
     } catch (e) {
       return false;
@@ -559,10 +558,9 @@ $validUntil: UInt32,$scalar: String!, $field: String!) {
       store.walletConnectService?.emitAccountsChanged(pubKey);
       store.assets!.setAssetsLoading(true);
       webApi.assets.fetchAllTokenAssets();
-      webApi.assets.fetchTransactions(pubKey);
       webApi.assets.fetchPendingTransactions(pubKey);
       webApi.assets.fetchPendingTransactions(pubKey);
-      webApi.assets.fetchZkTransactions(pubKey);
+      webApi.assets.fetchFullTransactions(pubKey);
       return true;
     } catch (e) {
       return false;